Move chrome_content_browser_client_unittest.cc to core_unit_tests
This file already lived directly in //chrome/browser but was only ever
compiled as part of the monolithic chrome/test:unit_tests target. Move
it into the existing core_unit_tests target instead.

Also fixes a sign-compare warning in the file that the old target's
config happened to suppress.

Bug: 353332589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8389560
Commit-Queue: Orko Garai <orko@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1697505}

+++ chrome/browser/chrome_content_browser_client_unittest.cc
@@ -647,7 +647,8 @@ TEST_F(ChromeContentBrowserClientTest,
       net::ERR_PROXY_AUTH_REQUESTED);
 
   ASSERT_TRUE(info);
-  EXPECT_EQ(info->resource_id, IDR_ENTERPRISE_PROXY_ERROR_PAGE_HTML);
+  EXPECT_EQ(info->resource_id,
+            static_cast<uint32_t>(IDR_ENTERPRISE_PROXY_ERROR_PAGE_HTML));
   auto override_param = info->alternative_error_page_params.FindBool(
       error_page::kOverrideErrorPage);
   ASSERT_TRUE(override_param.has_value());
